+++ src/test/fixtures/tabs-python-example.py
@@ -1,0 +1,4 @@
+import logging
+
+	def fizzBuzz
+		logging.info("	FizzBuzz")